Eliminate duplicate constants across frontend/backend - consolidate to shared
