+++ shared/user.interface.ts
@@ -12,6 +12,41 @@ export const PROFICIENCY_LEVELS = [
 
 export type ProficiencyLevel = typeof PROFICIENCY_LEVELS[number];
 
+/**
+ * Gender options for user profiles.
+ */
+export const GENDER_OPTIONS = [
+    'Male',
+    'Female',
+    'Non-binary',
+    'Prefer not to say'
+] as const;
+
+export type Gender = typeof GENDER_OPTIONS[number];
+
+/**
+ * Supported languages for language exchange platform.
+ */
+export const SUPPORTED_LANGUAGES = [
+    'English', 'Spanish', 'French', 'German', 'Italian', 'Portuguese',
+    'Russian', 'Chinese', 'Japanese', 'Korean', 'Arabic', 'Hindi',
+    'Dutch', 'Swedish', 'Norwegian', 'Danish', 'Finnish', 'Polish'
+] as const;
+
+export type SupportedLanguage = typeof SUPPORTED_LANGUAGES[number];
+
+/**
+ * Supported countries for user location selection.
+ */
+export const SUPPORTED_COUNTRIES = [
+    'United States', 'Canada', 'United Kingdom', 'Germany', 'France',
+    'Spain', 'Italy', 'Netherlands', 'Sweden', 'Norway', 'Denmark',
+    'Finland', 'Poland', 'Russia', 'China', 'Japan', 'South Korea',
+    'Australia', 'New Zealand', 'Brazil', 'Mexico', 'Argentina'
+] as const;
+
+export type SupportedCountry = typeof SUPPORTED_COUNTRIES[number];
+
 export interface IUser {
     // MongoDB has its own ID type. In the backend user.ts file, we are extending this file and the MongoDB document which have different ID types. THis creates an error. Research suggests using a new feature of Mongo DB (HydratedDocument). 
      _id: string;
@@ -32,11 +67,11 @@ export interface IUser {
     practicingLanguage: {
         language: string;
         proficiency: ProficiencyLevel;
-    };
-    country: string;
-    city: string;
-    gender: 'Male' | 'Female' | 'Non-binary' | 'Prefer not to say';
-    age: number;
+        };
+        country: SupportedCountry;
+        city: string;
+        gender: Gender;
+        age: number;
     };
 }
 
@@ -48,14 +83,14 @@ export interface IUserRegistrationRequest {
     familyName: string;
     bio: string;
     profileOptions: {
-        nativeLanguage: string;
+        nativeLanguage: SupportedLanguage;
         practicingLanguage: {
-            language: string;
+            language: SupportedLanguage;
             proficiency: ProficiencyLevel;
         };
-        country: string;
+        country: SupportedCountry;
         city: string;
-        gender: 'Male' | 'Female' | 'Non-binary' | 'Prefer not to say';
+        gender: Gender;
         age: number;
     };
 }
